# reports/rag_ingestion_auth_review/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatHistoryResponse": {
        "properties": {
          "latest_response": {
            "additionalProperties": true,
            "title": "Latest Response",
            "type": "object"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "session_id": {
            "default": "",
            "title": "Session Id",
            "type": "string"
          },
          "user_id": {
            "default": "demo_user",
            "title": "User Id",
            "type": "string"
          }
        },
        "title": "ChatHistoryResponse",
        "type": "object"
      },
      "ChatRequest": {
        "properties": {
          "channel": {
            "default": "test",
            "title": "Channel",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "default": "demo_user",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "answer_basis": {
            "default": "",
            "title": "Answer Basis",
            "type": "string"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Citations",
            "type": "array"
          },
          "confidence_level": {
            "default": "medium",
            "title": "Confidence Level",
            "type": "string"
          },
          "confidence_score": {
            "title": "Confidence Score",
            "type": "number"
          },
          "context_used": {
            "additionalProperties": true,
            "title": "Context Used",
            "type": "object"
          },
          "conversation_status": {
            "default": "pending_agent_review",
            "title": "Conversation Status",
            "type": "string"
          },
          "decision_trace": {
            "additionalProperties": true,
            "title": "Decision Trace",
            "type": "object"
          },
          "evaluation_metrics": {
            "additionalProperties": true,
            "title": "Evaluation Metrics",
            "type": "object"
          },
          "evidence_citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence Citations",
            "type": "array"
          },
          "expected_evidence_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Expected Evidence Keywords",
            "type": "array"
          },
          "expected_intent": {
            "default": "",
            "title": "Expected Intent",
            "type": "string"
          },
          "final_prompt": {
            "title": "Final Prompt",
            "type": "string"
          },
          "forbidden_keyword_hits": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Keyword Hits",
            "type": "array"
          },
          "forbidden_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Keywords",
            "type": "array"
          },
          "full_trace": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Full Trace",
            "type": "array"
          },
          "handoff_ticket": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handoff Ticket"
          },
          "human_review_reason": {
            "default": "v1 默认客服确认后发送",
            "title": "Human Review Reason",
            "type": "string"
          },
          "intent_analysis": {
            "additionalProperties": true,
            "title": "Intent Analysis",
            "type": "object"
          },
          "issue_type": {
            "default": "",
            "title": "Issue Type",
            "type": "string"
          },
          "manual_judgment": {
            "additionalProperties": true,
            "title": "Manual Judgment",
            "type": "object"
          },
          "matched_evidence_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Matched Evidence Keywords",
            "type": "array"
          },
          "memory_snapshot": {
            "additionalProperties": true,
            "title": "Memory Snapshot",
            "type": "object"
          },
          "missing_evidence_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Missing Evidence Keywords",
            "type": "array"
          },
          "mixed_supporting_intent": {
            "default": false,
            "title": "Mixed Supporting Intent",
            "type": "boolean"
          },
          "need_human_review": {
            "default": true,
            "title": "Need Human Review",
            "type": "boolean"
          },
          "needs_manual_review": {
            "default": false,
            "title": "Needs Manual Review",
            "type": "boolean"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "prompt_context_items": {
            "items": {
              "$ref": "#/components/schemas/PromptContextItemResponse"
            },
            "title": "Prompt Context Items",
            "type": "array"
          },
          "prompt_version": {
            "default": "",
            "title": "Prompt Version",
            "type": "string"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "request_id": {
            "default": "",
            "title": "Request Id",
            "type": "string"
          },
          "retrieved_documents": {
            "items": {
              "type": "string"
            },
            "title": "Retrieved Documents",
            "type": "array"
          },
          "retrieved_items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Retrieved Items",
            "type": "array"
          },
          "risk_level": {
            "default": "low",
            "title": "Risk Level",
            "type": "string"
          },
          "risky_promises": {
            "items": {
              "type": "string"
            },
            "title": "Risky Promises",
            "type": "array"
          },
          "safety_status": {
            "additionalProperties": true,
            "title": "Safety Status",
            "type": "object"
          },
          "session_id": {
            "default": "",
            "title": "Session Id",
            "type": "string"
          },
          "suggested_layer": {
            "default": "",
            "title": "Suggested Layer",
            "type": "string"
          },
          "token_usage": {
            "additionalProperties": true,
            "title": "Token Usage",
            "type": "object"
          },
          "tool_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tool Results",
            "type": "array"
          },
          "trace": {
            "$ref": "#/components/schemas/ChatTrace"
          },
          "used_primary_evidence": {
            "default": false,
            "title": "Used Primary Evidence",
            "type": "boolean"
          },
          "user_id": {
            "default": "demo_user",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "reply",
          "confidence_score",
          "final_prompt",
          "retrieved_documents",
          "trace"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ChatReviewActionRequest": {
        "properties": {
          "action": {
            "enum": [
              "accepted",
              "edited_and_sent",
              "human_handoff",
              "marked_bad_case"
            ],
            "title": "Action",
            "type": "string"
          },
          "final_reply": {
            "default": "",
            "title": "Final Reply",
            "type": "string"
          },
          "operator_id": {
            "default": "demo_agent",
            "title": "Operator Id",
            "type": "string"
          },
          "operator_role": {
            "default": "agent",
            "title": "Operator Role",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "action"
        ],
        "title": "ChatReviewActionRequest",
        "type": "object"
      },
      "ChatReviewActionResponse": {
        "properties": {
          "action": {
            "enum": [
              "accepted",
              "edited_and_sent",
              "human_handoff",
              "marked_bad_case"
            ],
            "title": "Action",
            "type": "string"
          },
          "audit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audit Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "final_reply": {
            "default": "",
            "title": "Final Reply",
            "type": "string"
          },
          "handoff_ticket": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handoff Ticket"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "saved": {
            "default": true,
            "title": "Saved",
            "type": "boolean"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "session_id",
          "user_id",
          "action",
          "status",
          "created_at"
        ],
        "title": "ChatReviewActionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "身份只来自 Authorization: Bearer <JWT>。X-User-Role / X-Operator-Id 已失效，仅在日志中记录，不参与任何授权判定。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LLM Customer Service API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat/history": {
      "get": {
        "operationId": "get_chat_history_chat_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "demo_user",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chat History",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/prompt": {
      "post": {
        "operationId": "generate_answer_chat_prompt_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Answer",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/review-action": {
      "post": {
        "operationId": "review_chat_action_chat_review_action_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatReviewActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReviewActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Chat Action",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/examples/by-category": {
      "get": {
        "operationId": "examples_examples_by_category_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamplesByCategoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Category not found"
                }
              }
            },
            "description": "Category not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Examples",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/categories": {
      "get": {
        "operationId": "categories_examples_categories_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Categories",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/search": {
      "post": {
        "operationId": "search_examples_api_examples_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchExamplesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchExamplesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Examples Api",
        "tags": [
          "examples"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}